# Conflicts:
#	src/org/graphstream/algorithm/APSP.java
#	src/org/graphstream/algorithm/PageRank.java
#	src/org/graphstream/algorithm/Toolkit.java
#	src/org/graphstream/algorithm/coloring/WelshPowell.java
#	src/org/graphstream/algorithm/flow/FordFulkersonAlgorithm.java
#	src/org/graphstream/algorithm/generator/LifeGenerator.java
#	src/org/graphstream/algorithm/generator/URLGenerator.java
#	src/org/graphstream/algorithm/measure/ConnectivityMeasure.java
#	src/org/graphstream/algorithm/measure/SurpriseMeasure.java
#	src/org/graphstream/algorithm/networksimplex/NetworkSimplex.java
#	src/org/graphstream/algorithm/randomWalk/RandomWalk.java
#	src/org/graphstream/ui/layout/HierarchicalLayout.java

+++ src/org/graphstream/algorithm/generator/ChvatalGenerator.java
@@ -79,7 +79,7 @@ public class ChvatalGenerator extends BaseGenerator {
 			{ false, false, false, false, false, false, false, false, false,
 					false, false, false }, };
 
-	private static final double[][] coordinates = { { 0, 0 }, { 5, 0 },
+	private static final int[][] coordinates = { { 0, 0 }, { 5, 0 },
 			{ 5, 5 }, { 0, 5 }, { 1, 2 }, { 2, 1 }, { 3, 1 }, { 4, 2 },
 			{ 4, 3 }, { 3, 4 }, { 2, 4 }, { 1, 3 } };
